refactoring approximation toolbox out of erf meta-function
Details:
*     + Creating new module core.approximation to host
*       the toolbox to ease approximation building
